+++ src/Shipment/Contracts/ShipmentStatus.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ShipmentStatus interface.
+ *
+ * @copyright   Copyright (c) 2019 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2019-02-16
+ *
+ */
+
+namespace Vanilo\Shipment\Contracts;
+
+interface ShipmentStatus
+{
+}