+++ src/Shipment/Models/ShipmentProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ShipmentProxy class.
+ *
+ * @copyright   Copyright (c) 2019 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2019-02-16
+ *
+ */
+
+namespace Vanilo\Shipment\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class ShipmentProxy extends ModelProxy
+{
+}