+++ src/Shipment/Tests/TestCase.php
@@ -1,0 +1,80 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the base TestCase class.
+ *
+ * @copyright   Copyright (c) 2019 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2019-02-16
+ *
+ */
+
+namespace Vanilo\Shipment\Tests;
+
+use Konekt\Address\Providers\ModuleServiceProvider as AddressModule;
+use Konekt\Concord\ConcordServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+use Vanilo\Shipment\Providers\ModuleServiceProvider as ShipmentModule;
+
+abstract class TestCase extends Orchestra
+{
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        $this->setUpDatabase($this->app);
+    }
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     *
+     * @return array
+     */
+    protected function getPackageProviders($app)
+    {
+        return [
+            ConcordServiceProvider::class
+        ];
+    }
+
+    /**
+     * Set up the environment.
+     *
+     * @param \Illuminate\Foundation\Application $app
+     */
+    protected function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('database.default', 'sqlite');
+        $app['config']->set('database.connections.sqlite', [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+        ]);
+    }
+
+    /**
+     * Set up the database.
+     *
+     * @param \Illuminate\Foundation\Application $app
+     */
+    protected function setUpDatabase($app)
+    {
+        \Artisan::call('migrate', ['--force' => true]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function resolveApplicationConfiguration($app)
+    {
+        parent::resolveApplicationConfiguration($app);
+
+        $app['config']->set('concord.modules', [
+            AddressModule::class,
+            ShipmentModule::class
+        ]);
+    }
+}