+++ src/Shipment/resources/manifest.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'name' => 'Vanilo Shipment Module',
+    'version' => '3.0-dev'
+];